Pure JS: add warnings for unhandled channel options
